M7/fix release

See merge request COMP9242/25T3/group-19!55

+++ projects/aos/sos/src/syscall_handler/sys_close.h
@@ -13,4 +13,4 @@
  *
  * @return 0 on success, or -1 on failure.
  */
-int handle_sos_close(size_t fd);
+int handle_sos_close(int fd);
